ci: publish coverage report to GitHub Pages

+++ lib/scripts/push-coverage-report.sh
@@ -1,15 +1,19 @@
 #!/bin/bash
 
-git config --global user.email 'engzerjun@gmail.com'
-git config --global user.name 'juneezee'
-git remote set-url origin https://acb16zje:StWb33dF6ygoCLxNwkoR@git.shefcompsci.org.uk/com4525-2019-20/team07/project.git
+if [ "$GH_PAGES" == "true" ]; then
+  git config --global user.email 'engzerjun@gmail.com'
+  git config --global user.name 'juneezee'
+  git remote remove origin
+  git remote add origin https://Juneezee:$GITHUB_TOKEN@github.com/Juneezee/project.git
 
-if git show-ref --quiet refs/heads/coverage-report; then
-  git branch -D coverage-report
+  if git show-ref --quiet refs/heads/gh-pages; then
+    git branch -D gh-pages
+  fi
+
+  git checkout --orphan gh-pages
+  git rm -rf .
+  mv coverage/* .
+  git add index.html assets/
+  git commit -m "docs: coverage report $(date +"%d-%b-%Y %R")"
+  git push -f origin gh-pages
 fi
-
-git checkout --orphan coverage-report
-git rm -rf .
-git add coverage/
-git commit -m "docs: coverage report $(date +"%d-%b-%Y %R")"
-git push -f origin coverage-report
